1.0
I used this plugin as a base but ended up changing a lot as I was learning. Should've setup the repository first but now it finished so we get one giant commit
also i made this a fabric mod at first oooops so there might be some remnants of that. good news is could could make it back into a fabric mod really easily! Not sure why youd want to but you can!